update model five day

+++ app/src/main/java/minhnq/gvn/com/openweathermap/model/Main.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName;
 public class Main {
     @SerializedName("temp")
     @Expose
-    public int temp;
+    public double temp;
 
     @SerializedName("temp_min")
     @Expose
